feat!: added ColorUtils class & Colorable interface
