+++ src/main/kotlin/viewmodel/Colorable.kt
@@ -1,0 +1,7 @@
+package viewmodel
+
+import androidx.compose.ui.graphics.Color
+
+interface Colorable {
+	var color: Color
+}
